code review: 코드리뷰 반영. 커스텀 훅 분리 및 useCallback을 통한 최적화
