+++ src/shared/constant/rate.ts
@@ -1,0 +1,1 @@
+export const RATES: number[] = [1, 2, 3, 4, 5];
